Refactor system utils
