+++ packages/azure-services/src/register-azure-services-to-container.ts
@@ -96,10 +96,14 @@ function setupSingletonQueueServiceURLProvider(container: interfaces.Container):
 
 function setupSingletonCosmosClientProvider(container: interfaces.Container): void {
     IoC.setupSingletonProvider<CosmosClient>(iocTypeNames.CosmosClientProvider, container, async context => {
-        const secretProvider = context.container.get(SecretProvider);
-        const cosmosDbUrl = await secretProvider.getSecret(secretNames.cosmosDbUrl);
-        const cosmosDbKey = await secretProvider.getSecret(secretNames.cosmosDbKey);
+        if (process.env.COSMOS_DB_URL !== undefined && process.env.COSMOS_DB_KEY !== undefined) {
+            return new CosmosClient({ endpoint: process.env.COSMOS_DB_URL, auth: { masterKey: process.env.COSMOS_DB_KEY } });
+        } else {
+            const secretProvider = context.container.get(SecretProvider);
+            const cosmosDbUrl = await secretProvider.getSecret(secretNames.cosmosDbUrl);
+            const cosmosDbKey = await secretProvider.getSecret(secretNames.cosmosDbKey);
 
-        return new CosmosClient({ endpoint: cosmosDbUrl, auth: { masterKey: cosmosDbKey } });
+            return new CosmosClient({ endpoint: cosmosDbUrl, auth: { masterKey: cosmosDbKey } });
+        }
     });
 }
