+++ packages/common/src/system/system.ts
@@ -2,9 +2,6 @@
 // Licensed under the MIT License.
 import * as crypto from 'crypto';
 import { isNil } from 'lodash';
-import * as nodeUrl from 'url';
-// @ts-ignore
-import * as uuid from 'uuid-with-v6';
 
 export namespace System {
     export function createInstanceIfNil<T>(instance: T, factory: () => T): T {
@@ -40,35 +37,5 @@ export namespace System {
             .randomBytes(bytes)
             .toString('hex')
             .substr(0, length);
-    }
-
-    export function createGuid(): string {
-        // tslint:disable-next-line: no-unsafe-any
-        return uuid.v6();
-    }
-
-    export function getGuidTimestamp(guid: string): Date {
-        if ((guid.length === 36 && guid.substr(14, 1) !== '6') || (guid.length === 32 && guid.substr(12, 1) !== '6')) {
-            throw new Error('Only version 6 of UUID is supported.');
-        }
-
-        const timestampMilliseconds = BigInt(`0x${guid.substr(0, 8)}${guid.substr(9, 4)}${guid.substr(15, 3)}`) / 10000n;
-        const baseDate = BigInt(Math.abs(new Date(Date.UTC(1582, 9, 15)).valueOf()));
-
-        return new Date(Number(timestampMilliseconds - baseDate));
-    }
-
-    export function tryParseUrlString(url: string, absoluteUrlOnly: boolean = true): nodeUrl.Url {
-        const absoluteUrlRegEx = /^(?:[a-z]+:)?\/\//i;
-        try {
-            const urlParsed = nodeUrl.parse(url);
-            if (absoluteUrlOnly && urlParsed.href.match(absoluteUrlRegEx) === null) {
-                return undefined;
-            }
-
-            return urlParsed;
-        } catch (error) {
-            return undefined;
-        }
     }
 }
